lift currency prefix logic up component tree
