+++ frontend/src/metabase/query_builder/components/filters/pickers/DefaultPicker.jsx
@@ -13,7 +13,10 @@ import FieldValuesWidget from "metabase/components/FieldValuesWidget";
 import {
   getFilterArgumentFormatOptions,
   isFuzzyOperator,
+  isCurrency,
 } from "metabase/lib/schema_metadata";
+
+import { getCurrencySymbol } from "metabase/lib/formatting";
 
 import {
   BetweenLayoutContainer,
@@ -61,6 +64,18 @@ export default function DefaultPicker({
   const isBetweenLayout =
     operator.name === "between" && operatorFields.length === 2;
 
+  const extractFieldSettings = field => {
+    const fieldId = field?.id;
+    const fieldMetadata = field?.metadata?.fields[fieldId];
+    return fieldMetadata?.settings;
+  };
+
+  const fieldSettings = extractFieldSettings(field);
+  const currencyPrefix =
+    isCurrency(field) && fieldSettings?.currency
+      ? getCurrencySymbol(fieldSettings.currency)
+      : null;
+
   const fieldWidgets = operatorFields
     .map((operatorField, index) => {
       let values, onValuesChange;
@@ -102,6 +117,7 @@ export default function DefaultPicker({
             multi={operator.multi}
             placeholder={placeholder}
             fields={underlyingField ? [underlyingField] : []}
+            prefix={currencyPrefix}
             disablePKRemappingForSearch={true}
             autoFocus={index === 0}
             alwaysShowOptions={operator.fields.length === 1}
@@ -132,6 +148,7 @@ export default function DefaultPicker({
             values={values}
             onValuesChange={onValuesChange}
             placeholder={placeholder}
+            prefix={currencyPrefix}
             multi={operator.multi}
             onCommit={onCommit}
           />
