+++ frontend/src/metabase/query_builder/components/filters/pickers/NumberPicker.jsx
@@ -3,8 +3,6 @@ import React, { Component } from "react";
 import PropTypes from "prop-types";
 import { t } from "ttag";
 import TextPicker from "./TextPicker";
-import { isCurrency } from "metabase/lib/schema_metadata";
-import { getCurrencySymbol } from "metabase/lib/formatting";
 
 export default class NumberPicker extends Component {
   constructor(props) {
@@ -53,18 +51,11 @@ export default class NumberPicker extends Component {
 
   render() {
     const values = this.state.stringValues.slice(0, this.props.values.length);
-
-    const fieldSettings = this.extractFieldSettings(this.props.field);
-    const currencyPrefix =
-      isCurrency(this.props.field) && fieldSettings?.currency
-        ? getCurrencySymbol(fieldSettings.currency)
-        : null;
-
     return (
       <TextPicker
         {...this.props}
         isSingleLine
-        prefix={currencyPrefix}
+        prefix={this.props.prefix}
         values={values}
         validations={this.state.validations}
         onValuesChange={values => this.onValuesChange(values)}
